vk: rename vk_textures to r_textures
