+++ ref/vk/infotool.c
@@ -1,7 +1,7 @@
 #include "camera.h"
 #include "vk_math.h"
 #include "vk_common.h"
-#include "vk_textures.h"
+#include "r_textures.h"
 #include "vk_brush.h"
 #include "vk_light.h"
