+++ ref/vk/r_textures.c
@@ -1,4 +1,4 @@
-#include "vk_textures.h"
+#include "r_textures.h"
 
 #include "vk_common.h"
 #include "vk_core.h"
