+++ ref/vk/vk_beams.c
@@ -3,7 +3,7 @@
 #include "camera.h"
 #include "vk_render.h"
 #include "vk_geometry.h"
-#include "vk_textures.h"
+#include "r_textures.h"
 #include "vk_sprite.h"
 #include "vk_scene.h"
 #include "vk_math.h"
