+++ ref/vk/vk_lightmap.c
@@ -1,6 +1,6 @@
 #include "vk_lightmap.h"
 #include "vk_common.h"
-#include "vk_textures.h"
+#include "r_textures.h"
 #include "vk_cvar.h"
 
 #include "com_strings.h"
