+++ ref/vk/vk_ray_model.c
@@ -1,7 +1,7 @@
 #include "vk_ray_internal.h"
 
 #include "vk_rtx.h"
-#include "vk_textures.h"
+#include "r_textures.h"
 #include "vk_materials.h"
 #include "vk_geometry.h"
 #include "vk_render.h"
