+++ ref/vk/vk_rmain.c
@@ -1,7 +1,7 @@
 #include "vk_core.h"
 #include "vk_cvar.h"
 #include "vk_common.h"
-#include "vk_textures.h"
+#include "r_textures.h"
 #include "vk_renderstate.h"
 #include "vk_overlay.h"
 #include "vk_scene.h"
