+++ ref/vk/vk_rtx.c
@@ -14,7 +14,7 @@
 #include "vk_pipeline.h"
 #include "vk_ray_internal.h"
 #include "vk_staging.h"
-#include "vk_textures.h"
+#include "r_textures.h"
 #include "vk_combuf.h"
 #include "vk_logs.h"
